refactor(logic): split world_market_phase into part-of concern files (#3290)
Phase-0 prerequisite slice for the colonizethis_logic package-split epic.
Decomposes the 914-line turn/phases/world_market_phase.dart (the largest
remaining >500-line file) into cohesive `part of` files, each well under
the Phase-0 500-line target, with zero behavior change.

Refs #3290
